+++ crates/peridot-cli/src/tests.rs
@@ -20,10 +20,11 @@ use super::run_loop::{
 use super::run_output::*;
 use super::run_state::*;
 use super::session_router::SessionRouter;
+use super::tui_session_export::render_session_export_text;
 use super::{
     AskUserPending, apply_session_command, context_top_report, delete_persisted_session,
-    hydrate_persisted_sessions, render_session_export_text, restore_latest_tui_state_from_disk,
-    restore_tui_state_from_disk, scan_and_suspend_running_sessions,
+    hydrate_persisted_sessions, restore_latest_tui_state_from_disk, restore_tui_state_from_disk,
+    scan_and_suspend_running_sessions,
 };
 
 #[test]